Fixes issues with Choice#getSelection in Firefox.
window.getSelection, for whatever reason, always had a range in
Firefox (even when nothing was selected, in which case the start
and end points were document.body), which caused some issues when
trying to save the selection; although the focus was clearly not
the root element, Choice#getSelection was not returning false.

The fix for this is to check that the root element is indeed focussed;
the downside is that the root element has to be focusable, which
restricts Choice's use cases to contentEditable elements.